Now the Database interface has the ServerQueries.

+++ client/src/main/java/net/ddns/mindustry/database/client/ServerQueries.java
@@ -1,10 +1,9 @@
 package net.ddns.mindustry.database.client;
 
 import net.ddns.mindustry.database.schema.tables.pojos.Server;
-import org.jooq.types.UShort;
 import java.util.Optional;
 
 public interface ServerQueries {
 
-    Optional<Server> find(String ip, UShort port);
+    Optional<Server> find(String ip, int port);
 }
